perf(web): keep scam/warranty libs out of vendor-app chunk (#3539)
Move the reactive scam-check and warranty-reminder injection effects out of
NotificationsContext (which lives in src/contexts/ and is hoisted into the
budget-constrained vendor-app shared chunk) into a new NotificationInjectors
component under components/notifications/. This keeps the heavier scam-alerts
and warranty libraries in the authenticated app graph -- where they lived
before the store was extracted -- restoring vendor-app below the lazy-chunk
gzip budget. The context module now only owns the shared useNotifications store.
